refactor(evm): keep precompile registration inline
